Modify signature of convolve slightly.
I have no idea whether these functions should
modify in-place, or return a new one.

The standard library usually has a convention where
their non-in-place modifying algorithms are suffixed
with _copy, i.e. convolve_copy(). We might decide to
do something like that. This is a case where we'll
want to think a bit to get the defaults right.